pyproject.toml
[project]
name = "image-dithering-sketch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
]
